backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
weasyprint==60.2
pdfplumber==0.11.0
pytesseract==0.3.10
Pillow==10.2.0
python-dotenv==1.0.0
# Auth dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
